chore: introduce useSubscriptions directory
