+++ pages/manage.tsx
@@ -1,24 +1,7 @@
-import { createContext, Dispatch, SetStateAction } from "react";
 import type { NextPage } from "next";
 import { SubscriptionForm } from "components/SubscriptionForm";
 import { SubscriptionList } from "components/SubscriptionList";
-import { Subscription, useSubscriptions } from "hooks/useSubscriptions";
-
-interface SubscriptionContextType {
-  subscriptions: Subscription[];
-  subscriptionsLoading: boolean;
-  subscriptionError: string;
-  setSubscriptions: Dispatch<SetStateAction<Subscription[]>>;
-  fetchSubscriptions: () => void;
-}
-
-export const SubscriptionContext = createContext<SubscriptionContextType>({
-  subscriptions: [],
-  subscriptionsLoading: false,
-  subscriptionError: "",
-  setSubscriptions: () => {},
-  fetchSubscriptions: () => {},
-});
+import { SubscriptionContext, useSubscriptions } from "hooks/useSubscriptions";
 
 const Manage: NextPage = () => {
   const useSubscriptionsPayload = useSubscriptions();
